Split stubs for readability
